backend/requirements.txt
# Core Django Framework
Django==5.2.6

# Django REST Framework for API endpoints  
djangorestframework==3.16

# CORS handling for frontend communication
django-cors-headers==4.3.1

# Environment variable management
python-decouple==3.8

# Firebase Admin SDK for authentication
firebase-admin==6.2.0

# Google Cloud Client Libraries
google-cloud-storage==2.14.0
google-cloud-firestore==2.13.1
google-cloud-aiplatform==1.38.1

# Authentication and retry logic
google-auth==2.23.4
tenacity==8.2.3

# Utilities
pydantic==2.5.2
python-jose==3.3.0
python-magic==0.4.27
python-dateutil==2.8.2

# Text extraction libraries
PyPDF2==3.0.1
python-docx==1.1.0
chardet==5.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-mock==3.12.0
